+++ Widgets/Dialog.php
@@ -7,7 +7,6 @@ use exface\Core\Interfaces\Model\MetaAttributeInterface;
 use exface\Core\DataTypes\BooleanDataType;
 use exface\Core\Factories\WidgetFactory;
 use exface\Core\Interfaces\Widgets\iContainOtherWidgets;
-use exface\Core\CommonLogic\Constants\Icons;
 use exface\Core\CommonLogic\UxonObject;
 use exface\Core\Exceptions\Widgets\WidgetConfigurationError;
 use exface\Core\Interfaces\Widgets\iHaveHeader;
@@ -17,6 +16,7 @@ use exface\Core\Factories\DataSheetFactory;
 use exface\Core\Interfaces\DataSheets\DataSheetInterface;
 use exface\Core\Interfaces\Actions\iShowDialog;
 use exface\Core\Interfaces\Actions\iCallOtherActions;
+use exface\Core\Widgets\Traits\PopupTrait;
 
 /**
  * Dialogs are pop-up forms (i.e. windows), that can be moved and/or maximized.
@@ -39,12 +39,8 @@ use exface\Core\Interfaces\Actions\iCallOtherActions;
  */
 class Dialog extends Form implements iAmClosable, iHaveHeader
 {
-    private $hide_close_button = false;
-
-    private $close_button = null;
+    use PopupTrait;
     
-    private $close_button_action = null;
-
     private $maximizable = true;
 
     private $maximized = null;
@@ -55,6 +51,11 @@ class Dialog extends Form implements iAmClosable, iHaveHeader
     
     private $cacheable = true;
 
+    /**
+     * 
+     * {@inheritDoc}
+     * @see \exface\Core\Widgets\AbstractWidget::init()
+     */
     protected function init()
     {
         parent::init();
@@ -81,101 +82,6 @@ class Dialog extends Form implements iAmClosable, iHaveHeader
             }
         }
         return parent::addWidget($widget, $position);
-    }
-
-    /**
-     * If TRUE, the automatically generated close button for the dialog is not shown
-     *
-     * @return boolean
-     */
-    public function getHideCloseButton()
-    {
-        return $this->hide_close_button;
-    }
-
-    /**
-     * If set to TRUE, the automatically generated close button will not be shown in this dialog
-     * 
-     * @uxon-property hide_close_button
-     * @uxon-type boolean
-     * @uxon-default false
-     *
-     * @param boolean $value 
-     * @return Dialog           
-     */
-    public function setHideCloseButton($value)
-    {
-        $this->hide_close_button = $value;
-        if ($this->close_button !== null) {
-            $this->close_button->setHidden($value);
-        }
-        return $this;
-    }
-
-    /**
-     * Returns a special dialog button, that just closes the dialog without doing any other action
-     *
-     * @return \exface\Core\Widgets\DialogButton
-     */
-    public function getCloseButton()
-    {
-        if (! ($this->close_button instanceof DialogButton)) {
-            /* @var $btn DialogButton */
-            $btn = $this->createButton();
-            $btn->setCloseDialogAfterActionSucceeds(true);
-            $btn->setRefreshInput(false);
-            $btn->setIcon(Icons::TIMES);
-            $btn->setCaption($this->translate('WIDGET.DIALOG.CLOSE_BUTTON_CAPTION'));
-            $btn->setAlign(EXF_ALIGN_OPPOSITE);
-            if ($this->getHideCloseButton()) {
-                $btn->setHidden(true);
-            }
-            if ($this->hasCloseButtonAction()) {
-                $btn->setAction($this->getCloseButtonActionUxon());
-            }
-            $this->close_button = $btn;
-        }
-        return $this->close_button;
-    }
-    
-    /**
-     * Sets the action, that the close button will trigger.
-     * 
-     * @uxon-property close_button_action
-     * @uxon-type \exface\Core\CommonLogic\AbstractAction
-     * @uxon-template {"alias": ""}
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iDefineAction::setAction()
-     */
-    public function setCloseButtonAction($uxon) : Dialog
-    {
-        $this->close_button_action = $uxon;
-        if ($this->close_button !== null) {
-            $this->close_button->setAction($uxon);
-        }
-        return $this;
-    }
-    
-    /**
-     * Returns NULL or the UXON for the close button action
-     * 
-     * @return UxonObject|NULL
-     */
-    public function getCloseButtonActionUxon() : ?UxonObject
-    {
-        return $this->close_button_action;
-    }
-    
-    /**
-     * Returns true if the close button has an action uxon
-     * 
-     * @return boolean
-     */
-    public function hasCloseButtonAction() : bool
-    {
-        return $this->close_button_action !== null;
     }
     
     /**
@@ -353,6 +259,9 @@ class Dialog extends Form implements iAmClosable, iHaveHeader
         return $this;
     }
     
+    /**
+     * @return DialogHeader
+     */
     public function getHeader()
     {
         return $this->header;
@@ -470,53 +379,6 @@ class Dialog extends Form implements iAmClosable, iHaveHeader
     {
         $this->cacheable = $value;
         return $this;
-    }
-    
-    /**
-     * Returns the button that opened the dialog or NULL if not available
-     *
-     * @return iTriggerAction|NULL
-     */
-    public function getOpenButton() : ?iTriggerAction
-    {
-        if ($this->hasParent()) {
-            $parent = $this->getParent();
-            if ($parent instanceof iTriggerAction) {
-                return $parent;
-            }
-        }
-        return null;
-    }
-    
-    /**
-     * Returns the action that opened the dialog or NULL if not available
-     *
-     * @return iShowDialog|NULL
-     */
-    public function getOpenAction() : ?iShowDialog
-    {
-        if ($button = $this->getOpenButton()) {
-            if ($button->hasAction()) {
-                $action = $button->getAction();
-                switch (true) {
-                    // If the action shows a dialog, that's it
-                    case $action instanceof iShowDialog:
-                        return $action;
-                        // If the action calls other actions, see if one of them fits: that is, it shows exactly
-                        // our dialog. It is importantch to check for the dialog id as CallAction action might
-                        // have multiple actions calling each their own dialog!
-                    case $action instanceof iCallOtherActions:
-                            $thisDialogId = $this->getId();
-                            foreach ($action->getActionsRecursive() as $innerAction) {
-                                if (($innerAction instanceof iShowDialog) && $innerAction->getDialogWidget()->getId() === $thisDialogId) {
-                                    return $innerAction;
-                                }
-                            }
-                            break;
-                }
-            }
-        }
-        return null;
     }
     
     /**
